Move a venue-related method into the appropriate partial class (#2104)
