+++ src/Dfc.CourseDirectory.Core/DataManagement/FileUploadProcessor.cs
@@ -7,10 +7,8 @@ using System.Reactive.Linq;
 using System.Threading.Tasks;
 using Azure.Storage.Blobs;
 using CsvHelper;
-using Dfc.CourseDirectory.Core.DataManagement.Schemas;
 using Dfc.CourseDirectory.Core.DataStore;
 using Dfc.CourseDirectory.Core.DataStore.Sql;
-using Dfc.CourseDirectory.Core.DataStore.Sql.Queries;
 
 namespace Dfc.CourseDirectory.Core.DataManagement
 {
@@ -197,48 +195,6 @@ namespace Dfc.CourseDirectory.Core.DataManagement
             {
                 throw new ArgumentException("Stream must be seekable.", nameof(stream));
             }
-        }
-
-        public async Task<bool> DeleteVenueUploadRowForProvider(Guid providerId, int rowNumber)
-        {
-            using (var dispatcher = _sqlQueryDispatcherFactory.CreateDispatcher())
-            {
-                var venueUpload = await dispatcher.ExecuteQuery(new GetLatestUnpublishedVenueUploadForProvider()
-                {
-                    ProviderId = providerId
-                });
-
-                var (existingRows, lastRowNumber) = await dispatcher.ExecuteQuery(new GetVenueUploadRows() { VenueUploadId = venueUpload.VenueUploadId });
-
-                var rowToDelete = existingRows.SingleOrDefault(x => x.RowNumber == rowNumber);
-                if (rowToDelete == null)
-                {
-                    return false;
-                }
-
-                if (!rowToDelete.IsDeletable)
-                {
-                    throw new InvalidStateException(InvalidStateReason.VenueUploadRowCannotBeDeleted);
-                }
-
-                var nonDeletedRows = existingRows.Where(x => x.RowNumber != rowNumber).ToArray();
-
-                var rowCollection =  new VenueDataUploadRowInfoCollection(
-                    lastRowNumber: lastRowNumber,
-                    nonDeletedRows
-                        .Where(r => r.RowNumber != rowNumber)
-                        .Select(r => new VenueDataUploadRowInfo(CsvVenueRow.FromModel(r), r.RowNumber, r.IsSupplementary)));
-
-                await ValidateVenueUploadFile(
-                    dispatcher,
-                    venueUpload.VenueUploadId,
-                    venueUpload.ProviderId,
-                    rowCollection);
-
-                await dispatcher.Commit();
-            }
-
-            return true;
         }
 
         protected internal enum FileMatchesSchemaResult
